+++ examples/09-exchange.ts
@@ -1,30 +1,14 @@
-import { DataUtils, PrimeDataUtils, graphqlEndpoints } from '../src';
+import { DataUtils } from '../src';
 import * as dotenv from 'dotenv';
-import { BigNumber, constants } from 'ethers';
 
 dotenv.config();
+const dataApiKey = 'eyJvcmciOiI2NTIzZjY5MzUwOTBmNzAwMDFiYjJkZWIiLCJpZCI6IjI4ZWJiMGQ5YTMxYjQ3MmY4NmU4MWY2YTVhYzBhMzE1IiwiaCI6Im11cm11cjEyOCJ9';
 
 async function main(): Promise<void> {
   // initializating Data service...
-  const dataService = new DataUtils('public-prime-testnet-key', graphqlEndpoints.QA);
-  const primeDataService = new PrimeDataUtils(process.env.DATA_API_KEY);
-  const exchangeSupportedAssets = await primeDataService.getExchangeSupportedAssets({ page: 1, limit: 100, account: '', chainId: Number(process.env.CHAIN_ID) });
+  const dataService = new DataUtils(dataApiKey);
+  const exchangeSupportedAssets = await dataService.getExchangeSupportedAssets({ page: 1, limit: 100, account: '', chainId: Number(process.env.CHAIN_ID) });
   console.log('\x1b[33m%s\x1b[0m', `Found exchange supported assets:`, exchangeSupportedAssets.items.length);
-
-  const fromTokenAddress = '0xe3818504c1b32bf1557b16c238b2e01fd3149c17';
-  const toTokenAddress = constants.AddressZero;
-  const fromAmount = '1000000000000000000';
-  const fromChainId = 1;
-
-  const offers = await dataService.getExchangeOffers({
-    fromAddress: '',
-    fromChainId,
-    fromTokenAddress,
-    toTokenAddress,
-    fromAmount: BigNumber.from(fromAmount),
-  });
-
-  console.log('\x1b[33m%s\x1b[0m', `Exchange offers:`, offers);
 }
 
 main()