+++ src/sdk/data/index.ts
@@ -1,4 +1,3 @@
 export * from './data.module';
 export * from './classes';
 export * from './constants';
-export * from './primeData.module';